# Conflicts:
#	Phi6 Alpha/Base.lproj/Main.storyboard

+++ Phi6 Alpha/GameScene.swift
@@ -112,6 +112,8 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
     var viewController: GameViewController!
     var alertMessage: Any?
     var adder = Adder()
+    var locked = false
+    var editable = true
     
     var shapeLayerRope = CAShapeLayer()
     var pendulumFulcrum: CGPoint!
@@ -221,7 +223,7 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
                         Singleton.shared.setPosition(position: touchLocation)
                         print(node.name!)
                         if let sprite = node as? SKSpriteNode {
-                            if sprite == phisphere {
+                            if sprite == phisphere && !locked {
                                 deleteSliders()
                                 phisphere.position = touchLocation
                                 adder.addInitSlider(scene: self)
@@ -327,15 +329,23 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
             if let touch = touches.first {
                 let touchLocation = touch.location(in: self)
                 let touchedWhere = nodes(at: touchLocation)
-                let firstNode = touchedWhere.first!
+                //let firstNode = touchedWhere.first!
+                
+                let firstNode: SKNode!
+                if let first = touchedWhere.first {
+                    firstNode = first
                 
                 if !touchedWhere.isEmpty {
                     for node in [firstNode] {
                         if let sprite = node as? SKSpriteNode {
                             if sprite == phisphere {
-                                phisphere.position = touchLocation
-                                adder.addInitSlider(scene: self)
-                                myNode = sprite
+                                if !locked {
+                                    phisphere.position = touchLocation
+                                }
+                                if editable {
+                                    adder.addInitSlider(scene: self)
+                                    myNode = sprite
+                                }
                             } else {
                                 for object in Singleton.shared.objects {
                                     if sprite == object {
@@ -402,6 +412,7 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
                             }
                         }
                     }
+                }
                 }
             }
         }
